feat: Update PPO configuration to include increased number of environment runners
Update the PPO configuration in `ppo.py` to increase the number of environment runners from 4 to 8. This change will improve the efficiency of the training process by allowing more environments to be run in parallel.

Also, update the environment name in `main.py` from "ReferenceModel-1-1" to "ReferenceModel-1-2". This change reflects the updated version of the reference model being used.

Lastly, in `reference_model_1_1.py`, add support for variable sensor range in the `ReferenceModel` class. The sensor range determines the size of the observation grid around each agent. This change allows for more flexibility in defining the observation space for the agents.

These changes enhance the performance and flexibility of the PPO algorithm and the reference model environment.